Adding some new flavors tests
